httpd-keda-autoscaling/scaling-tests/requirements.txt
behave==1.2.6
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
dateparser==1.2.0
fonttools==4.53.1
google-auth==2.34.0
httmock==1.4.0
idna==3.8
kiwisolver==1.4.7
kubernetes==30.1.0
matplotlib==3.9.2
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
parse==1.20.2
parse_type==0.6.3
pillow==10.4.0
prometheus-api-client==0.5.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
websocket-client==1.8.0
